** version check improved  ** setting to change url to profile version xml (change it manually in RSXSettings.xml)  ** partial x64 port (needs a recompiled libs and warning fixes)  ** added fav tabs icons and a fav dlg icon  ** some stuff in downloadmanager  ** fixed dropping msg in pm frame

+++ trunk/windows/SpyFrame.cpp
@@ -168,7 +168,7 @@ LRESULT SpyFrame::onSpeaker(UINT /*uMsg*/, WPARAM wParam, LPARAM lParam, BOOL& /
 		delete x;
 
 		ctrlStatus.SetText(2, (TSTRING(TOTAL) + Util::toStringW(total)).c_str());
-		ctrlStatus.SetText(4, (TSTRING(HITS) + Util::toStringW(ShareManager::getInstance()->getHits())).c_str());
+		ctrlStatus.SetText(4, (TSTRING(HITS) + Util::toStringW((size_t)ShareManager::getInstance()->getHits())).c_str());
 		double ratio = total > 0 ? ((double)ShareManager::getInstance()->getHits()) / (double)total : 0.0;
 		ctrlStatus.SetText(5, (TSTRING(HIT_RATIO) + Util::toStringW(ratio)).c_str());
 	} else if(wParam == TICK_AVG) {
